learn-api/requirements.txt
fastapi>=0.70.0
python-multipart
python-jose[cryptography]
passlib[bcrypt]~=1.7.4
celery~=5.2.3
uvicorn~=0.17.5
pyvips>=2.1.14
pydantic~=1.8.1
aiofiles~=0.6.0
mysql-connector-python
SQLAlchemy~=1.4.2
starlette>=0.16.0
pydantic[email]
shortuuid~=1.0.1
requests~=2.25.1
opencv-python==4.5.5.64
Shapely~=1.8.0
numpy>=1.21.2
imutils~=0.5.4
alembic~=1.7.1
minio~=7.1.0
pytest~=6.2.5
XlsxWriter>=3.0.1
pandas >= 1.3.4
black>=22.3.0
loguru>=0.6.0
sentry-sdk>=1.14.0
supertokens-python==0.12.8